# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Easy Boost API
  version: 1.0.0
  description: Production HTTP API. Authentication uses an HttpOnly SameSite cookie.
servers:
  - url: https://useboost.ru
security:
  - cookieAuth: []
paths:
...
  /api/v1/tg/start:
    post:
      security: []
      summary: Create a short-lived Telegram authorization code
      responses:
        '200':
          description: Authorization URL and opaque code
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/tg/check:
    get:
      security: []
      summary: Check and consume Telegram authorization code
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string, maxLength: 256 }
      responses:
        '200': { description: Pending or authenticated state }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/me:
    get:
      summary: Current user, base subscription, Premium entitlements, remaining voice minutes and server-projected feature flags
      responses:
        '200': { description: User profile; `features.adaptive_learning` is true only when the server rollout flag is enabled }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/payments/requests:
    post:
      summary: Create or return the current user's pending Premium Voice payment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product]
              properties:
                product: { type: string, enum: [premium_voice] }
              additionalProperties: false
      responses:
        '201': { description: New request created; no subscription or entitlement is self-granted }
        '200': { description: Existing pending request returned idempotently }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      summary: Read the current user's latest Premium Voice payment request and entitlement status
      parameters:
        - in: query
          name: product
          required: true
          schema: { type: string, enum: [premium_voice] }
      responses:
        '200': { description: Latest request or null plus current entitlement_active flag }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/admin/payment-requests/{requestId}/resolve:
    post:
      summary: Idempotently approve or reject a payment request as an administrator
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
              additionalProperties: false
      responses:
        '200': { description: Terminal request result. Premium approval atomically grants or extends the base subscription and bounded voice_tutor entitlement to the same end time; rejection grants nothing. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: Administrator role required; students cannot self-grant and administrators cannot approve their own request }
        '404': { description: Payment request not found }
  /api/v1/admin/payment-requests:
    get:
      summary: List the actionable Premium Voice payment queue for an administrator
      parameters:
        - in: query
          name: product
          required: true
          schema: { type: string, enum: [premium_voice] }
        - in: query
          name: status
          required: true
          schema: { type: string, enum: [new] }
      responses:
        '200': { description: Bounded request UUID, username, product and status rows used to resolve verified payments }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: Administrator role required }
  /api/v1/admin/users/{username}/entitlements/voice_tutor/revoke:
    post:
      summary: Idempotently revoke the active Voice Tutor entitlement without rewriting base subscription history
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        '200': { description: Whether an active entitlement was revoked }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: Administrator role required }
...
  /api/v1/admin/status:
    get:
      summary: Verify administrator access
      responses:
        '200': { description: Administrator access confirmed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Authenticated user does not have the admin role }
...
  /api/v1/adaptive-learning/goal:
    get:
      summary: Read the authenticated learner's current EGE goal
      responses:
        '200':
          description: Current goal or null
          content:
            application/json:
              schema:
                type: object
                required: [goal]
                properties:
                  goal:
                    allOf:
                      - $ref: '#/components/schemas/AdaptiveGoal'
                    nullable: true
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Adaptive Learning rollout is disabled; the route is not registered }
    put:
      summary: Idempotently save a current EGE target and refresh the evidence-backed preliminary profile
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetExam, targetScore, examDate, weeklyMinutes]
              properties:
                targetExam: { type: string, enum: [ege_english] }
                targetScore: { type: integer, minimum: 0, maximum: 100 }
                examDate: { type: string, format: date, description: Must be later than the current UTC date. }
                weeklyMinutes: { type: integer, minimum: 30, maximum: 2520, multipleOf: 5 }
              additionalProperties: false
      responses:
        '201': { description: Goal revision created and current profile returned }
        '200': { description: Exact idempotent replay; prior goal revision and current profile returned }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Idempotency key was already used with a different goal payload }
        '404': { description: Adaptive Learning rollout is disabled; the route is not registered }
  /api/v1/adaptive-learning/overview:
    get:
      summary: Read the current goal and an owner-bound profile bootstrapped from persisted attempts and Voice Tutor recoveries
      responses:
        '200':
          description: Goal plus transparent preliminary or established profile; no raw answer, essay, transcript or audio is copied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdaptiveOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Adaptive Learning rollout is disabled; the route is not registered }
  /api/v1/account/export:
    get:
      summary: Export the authenticated user's account, progress and attempt data
      responses:
        '200': { description: JSON export including adaptive goals/structured estimates, Voice Tutor sessions, recovery/repeat outcomes and bounded provider/model/prompt metadata; no raw audio, temporary captions, full Voice Tutor transcript, nonce hash or repeat answer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/account:
    delete:
      summary: Permanently delete the authenticated user and related data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmation]
              properties:
                confirmation: { type: string, enum: [DELETE] }
      responses:
        '200': { description: Account, Voice Tutor sessions/recoveries/repeats and pending/rejected rule reports are deleted; approved canonical cards survive without creator identity, and reviewer identity on retained cards is anonymized }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/privacy/consent:
    get:
      summary: Read the user's text and realtime speech-to-speech processing choices and current policy version
      responses:
        '200': { description: Consent state and current policy version }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      summary: Grant or withdraw text and realtime speech-to-speech processing consent; an outdated policy version does not authorize a Voice Tutor connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text_processing, voice_processing]
              properties:
                text_processing: { type: boolean }
                voice_processing: { type: boolean }
      responses:
        '200': { description: Updated consent state }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: eb_token
  schemas:
    AdaptiveGoal:
      type: object
      required: [id, targetExam, targetScore, examDate, weeklyMinutes, revision, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        targetExam: { type: string, enum: [ege_english] }
        targetScore: { type: integer, minimum: 0, maximum: 100 }
        examDate: { type: string, format: date }
        weeklyMinutes: { type: integer, minimum: 30, maximum: 2520, multipleOf: 5 }
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
    AdaptiveSkillEstimate:
      type: object
      required: [id, label, module, mastery, uncertainty, status, evidenceCount, effectiveEvidenceCount, independentEvidenceCount, evidenceQuality, lastObservedAt, dueState, explanationCode]
      properties:
        id: { type: string, maxLength: 120 }
        label: { type: string, maxLength: 160 }
        module: { type: string, enum: [grammar, vocabulary, reading, listening, writing, speaking] }
        mastery:
          type: integer
          minimum: 0
          maximum: 100
          description: Conservative estimate; without independent evidence it cannot exceed 49.
        uncertainty:
          type: integer
          minimum: 0
          maximum: 100
          description: Reduced only by independent unassisted or retention evidence.
        status:
          type: string
          enum: [unobserved, preliminary, established]
          description: A skill is established only after at least two independent observations.
        evidenceCount: { type: integer, minimum: 0 }
        effectiveEvidenceCount:
          type: integer
          minimum: 0
          description: Observations used in scoring after trust ceilings; raw evidenceCount remains auditable.
        independentEvidenceCount: { type: integer, minimum: 0 }
        evidenceQuality: { type: string, enum: [none, client_reported, assisted, independent, mixed] }
        lastObservedAt: { type: string, format: date-time, nullable: true }
        dueState: { type: string, enum: [not_due, due, overdue] }
        explanationCode: { type: string, maxLength: 40 }
      additionalProperties: false
    AdaptiveProfile:
      type: object
      required: [taxonomyVersion, weightingVersion, profileCalculationRevision, evidenceWatermarkVersion, evidenceObservedAt, evidenceSourceCount, preliminary, status, confidence, evidenceCount, independentEvidenceCount, assistedEvidenceCount, clientReportedEvidenceCount, independentModuleCount, establishedSkillCount, needsDiagnostic, explanationCodes, skills, modules]
      properties:
        taxonomyVersion: { type: string, enum: [ege-en-v1] }
        weightingVersion: { type: string, enum: [adaptive-evidence-v1] }
        profileCalculationRevision:
          type: integer
          minimum: 1
          description: Monotonic profile algorithm revision; an older calculation can never replace a newer one, while no revision may reduce evidenceSourceCount.
        evidenceWatermarkVersion: { type: string, enum: [adaptive-evidence-watermark-v1] }
        evidenceObservedAt:
          type: string
          format: date-time
          nullable: true
          description: Latest valid source-event timestamp used as the equal-count ordering tie-breaker.
        evidenceSourceCount:
          type: integer
          minimum: 0
          description: Append-only raw source-event count; lower counts never replace higher counts within one calculation revision and may include safely uncredited events.
        preliminary: { type: boolean }
        status: { type: string, enum: [preliminary, established] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        evidenceCount: { type: integer, minimum: 0 }
        independentEvidenceCount: { type: integer, minimum: 0 }
        assistedEvidenceCount: { type: integer, minimum: 0 }
        clientReportedEvidenceCount: { type: integer, minimum: 0 }
        independentModuleCount: { type: integer, minimum: 0, maximum: 6 }
        establishedSkillCount:
          type: integer
          minimum: 0
          maximum: 12
          description: Number of independently established taxonomy skills; the whole profile stays preliminary below 12.
        needsDiagnostic: { type: boolean }
        explanationCodes: { type: array, items: { type: string } }
        skills: { type: array, items: { $ref: '#/components/schemas/AdaptiveSkillEstimate' } }
        modules: { type: array, items: { type: object } }
      additionalProperties: false
    AdaptiveOverview:
      type: object
      required: [goal, profile]
      properties:
        goal:
          allOf:
            - $ref: '#/components/schemas/AdaptiveGoal'
          nullable: true
        profile: { $ref: '#/components/schemas/AdaptiveProfile' }
      additionalProperties: false
...
  responses:
    BadRequest: { description: Invalid request }
    Unauthorized: { description: Missing or expired session }
...
    RateLimited: { description: Rate limit exceeded }
...